backend/requirements.txt
# Temporal + LangGraph + ADO
temporalio>=1.9.0
httpx>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0
langchain-openai>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
fastapi>=0.115.0
uvicorn>=0.34.0
